Rename pauli.py -> pauli_gates.py
