+++ cirq/ops/pauli_gates.py
